requirements.txt
apache-airflow==2.8.1
praw==7.7.1
pandas==2.2.1
nltk==3.8.1
praw
